requirements.txt
Flask==3.0.3
spacy==3.8.5
psycopg2-binary==2.9.9
gunicorn==21.2.0
Jinja2==3.1.3
